Move the `DebugInterface` class inside the Go client
I moved the `DebugInterface` class into a separate package so that it 
could be imported from the outside.
