+++ clients/go/debugginginterface/debug_interface.go
@@ -1,4 +1,4 @@
-package main
+package debugginginterface
 
 import (
     "bufio"
