+++ clients/go/my_strategy.go
@@ -2,6 +2,7 @@ package main
 
 import (
     . "ai_cup_22/model"
+    . "ai_cup_22/debugginginterface"
 )
 
 type MyStrategy struct{}
